Move common security configs into spring-cloud-common-security
 - Move all the extended security configuration classes into a separate project `spring-cloud-common-security` so that other projects could re-use them
 - Update `spring-cloud-common-security` as a dependency
 - Update tests

Add customizable properties for authorization

 - Use the authenticated/permitted path properties in config
 - Refactor security configuration to support the existing `spring.cloud.dataflow.security` prefix for the security configuration properties
 - Update the tests

Use the actual properties/configuration classes

 - instead of subclassing
