+++ pom.xml
@@ -39,6 +39,8 @@
 		<!-- Note: Change ./spring-cloud-dataflow-server-local/pom.xml to match the spring-cloud.version -->
 		<spring-cloud.version>Dalston.SR2</spring-cloud.version>
 
+		<spring-cloud-security.version>1.0.0.BUILD-SNAPSHOT</spring-cloud-security.version>
+
 		<spring-analytics.version>1.1.3.RELEASE</spring-analytics.version>
 		<spring-batch-admin-manager.version>1.3.1.RELEASE</spring-batch-admin-manager.version>
 		<spring-shell.version>1.2.0.RELEASE</spring-shell.version>
@@ -123,6 +125,11 @@
 				<groupId>org.springframework.session</groupId>
 				<artifactId>spring-session</artifactId>
 				<version>${spring-session.version}</version>
+			</dependency>
+			<dependency>
+				<groupId>org.springframework.cloud</groupId>
+				<artifactId>spring-cloud-common-security</artifactId>
+				<version>1.0.0.BUILD-SNAPSHOT</version>
 			</dependency>
 			<dependency>
 				<groupId>org.springframework.cloud</groupId>
