--- spring-cloud-dataflow-server-core/src/main/java/org/springframework/cloud/dataflow/server/config/security/AuthorizationConfig.java
@@ -1,53 +1,0 @@
-/*
- * Copyright 2016-2017 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.springframework.cloud.dataflow.server.config.security;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.cloud.dataflow.core.DataFlowPropertyKeys;
-
-/**
- * Holds configuration for the authorization aspects of security.
- *
- * @author Eric Bottard
- * @author Gunnar Hillert
- */
-@ConfigurationProperties(prefix = DataFlowPropertyKeys.PREFIX + "security.authorization")
-public class AuthorizationConfig {
-
-	private boolean enabled = true;
-
-	private List<String> rules = new ArrayList<>();
-
-	public List<String> getRules() {
-		return rules;
-	}
-
-	public void setRules(List<String> rules) {
-		this.rules = rules;
-	}
-
-	public boolean isEnabled() {
-		return enabled;
-	}
-
-	public void setEnabled(boolean enabled) {
-		this.enabled = enabled;
-	}
-
-}
