--- spring-cloud-dataflow-server-core/src/main/java/org/springframework/cloud/dataflow/server/config/security/support/LdapSecurityProperties.java
@@ -1,126 +1,0 @@
-/*
- * Copyright 2016 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.springframework.cloud.dataflow.server.config.security.support;
-
-import java.net.URI;
-
-import javax.validation.constraints.NotNull;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.cloud.dataflow.core.DataFlowPropertyKeys;
-
-/**
- * Properties for the Ldap security aspects of an application.
- *
- * @author Gunnar Hillert
- * @since 1.1.0
- */
-@ConfigurationProperties(prefix = DataFlowPropertyKeys.PREFIX + "security.authentication.ldap")
-@LdapSecurityPropertiesValid
-public class LdapSecurityProperties {
-
-	@NotNull(message = "Provide a valid url to your Ldap server")
-	private URI url;
-
-	private String userDnPattern;
-
-	private String managerDn;
-
-	private String managerPassword;
-
-	private String userSearchBase = "";
-
-	private String userSearchFilter;
-
-	private String groupSearchFilter = "";
-
-	private String groupSearchBase = "";
-
-	private String groupRoleAttribute = "cn";
-
-	public URI getUrl() {
-		return url;
-	}
-
-	public void setUrl(URI url) {
-		this.url = url;
-	}
-
-	public String getUserDnPattern() {
-		return userDnPattern;
-	}
-
-	public void setUserDnPattern(String userDnPattern) {
-		this.userDnPattern = userDnPattern;
-	}
-
-	public String getManagerDn() {
-		return managerDn;
-	}
-
-	public void setManagerDn(String managerDn) {
-		this.managerDn = managerDn;
-	}
-
-	public String getManagerPassword() {
-		return managerPassword;
-	}
-
-	public void setManagerPassword(String managerPassword) {
-		this.managerPassword = managerPassword;
-	}
-
-	public String getUserSearchBase() {
-		return userSearchBase;
-	}
-
-	public void setUserSearchBase(String userSearchBase) {
-		this.userSearchBase = userSearchBase;
-	}
-
-	public String getUserSearchFilter() {
-		return userSearchFilter;
-	}
-
-	public void setUserSearchFilter(String userSearchFilter) {
-		this.userSearchFilter = userSearchFilter;
-	}
-
-	public String getGroupSearchFilter() {
-		return groupSearchFilter;
-	}
-
-	public void setGroupSearchFilter(String groupSearchFilter) {
-		this.groupSearchFilter = groupSearchFilter;
-	}
-
-	public String getGroupSearchBase() {
-		return groupSearchBase;
-	}
-
-	public void setGroupSearchBase(String groupSearchBase) {
-		this.groupSearchBase = groupSearchBase;
-	}
-
-	public String getGroupRoleAttribute() {
-		return groupRoleAttribute;
-	}
-
-	public void setGroupRoleAttribute(String groupRoleAttribute) {
-		this.groupRoleAttribute = groupRoleAttribute;
-	}
-
-}
