+++ spring-cloud-dataflow-server-core/src/main/java/org/springframework/cloud/dataflow/server/controller/security/SecurityController.java
@@ -18,7 +18,7 @@ package org.springframework.cloud.dataflow.server.controller.security;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cloud.dataflow.rest.resource.security.SecurityInfoResource;
-import org.springframework.cloud.dataflow.server.config.security.support.SecurityStateBean;
+import org.springframework.cloud.security.support.SecurityStateBean;
 import org.springframework.hateoas.ExposesResourceFor;
 import org.springframework.hateoas.mvc.ControllerLinkBuilder;
 import org.springframework.http.HttpStatus;
