+++ spring-cloud-dataflow-server-core/src/main/resources/META-INF/dataflow-server-defaults.yml
@@ -23,6 +23,7 @@ spring:
       security:
         authorization:
           enabled: true
+          permit-all-paths: "/authenticate,/security/info,/features,/assets/**"
           rules:
             # Metrics
